backend/azure_functions/requirements.txt
azure-functions
azure-cosmos
PyJWT